pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yt-transcribe"
version = "0.1.0"
description = "CLI tool to download YouTube subtitles with intelligent fallbacks and optional translation"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp>=2024.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
yt = "yt.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/yt"]

[tool.hatch.build.targets.sdist]
include = ["src/yt"]
